requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blis==0.4.1
cachetools==5.3.0
catalogue==1.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
editdistance==0.6.2
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.0/en_core_web_lg-2.3.0.tar.gz
executing==1.2.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mordecai==2.1.0
murmurhash==1.0.9
nest-asyncio==1.5.6
numpy==1.21.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
plac==1.1.3
platformdirs==3.1.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
#pywin32==305
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spacy==2.3.0
srsly==1.0.6
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
#tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==7.4.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
